feat(apple): handle scroll to top on profile page
